+++ cada/api.py
@@ -1,16 +1,21 @@
 # -*- coding: utf-8 -*-
-from flask import Blueprint, render_template, jsonify
+from flask import Blueprint, render_template, jsonify, json, url_for
 
 from cada.models import Advice
 from cada.search import search_advices
 
+import requests
 
 api = Blueprint('api', __name__)
 
 
 @api.route('/')
 def doc():
-    sample = Advice.objects.first()
+    sample = {
+        "advice": Advice.objects.first(),
+        "search": sample_search()
+    }
+
     return render_template('api.html', sample=sample)
 
 
@@ -27,6 +32,7 @@ def display(id):
     return jsonify(_serialize(advice))
 
 
+@api.app_template_filter()
 def _serialize(advice):
     return {
         'id': advice.id,
@@ -42,5 +48,16 @@ def _serialize(advice):
     }
 
 
+@api.app_template_filter()
+def pretty_json(value):
+    return json.dumps(obj=value, sort_keys=True, indent=4, separators=(',', ': '))
+
+
 def init_app(app):
     app.register_blueprint(api, url_prefix='/api')
+
+
+def sample_search():
+    r = requests.get(url_for('api.search', q='Paris', sort='session desc', page_size=3, _external=True))
+
+    return r.json()