+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## Current (in progress)
 
+- Python 3 [#28](https://github.com/etalab/cada.data.gouv.fr/pull/28)
 - Elastic Search 7.2 [#29](https://github.com/etalab/cada.data.gouv.fr/pull/29)
 - Mongo 4.1 [#29](https://github.com/etalab/cada.data.gouv.fr/pull/29)
-- Python 3 [#28](https://github.com/etalab/cada.data.gouv.fr/pull/28)
 - Improved pseudonymization [#32](https://github.com/etalab/cada.data.gouv.fr/pull/32)
+- Display sample of api calls [#33](https://github.com/etalab/cada.data.gouv.fr/pull/33)
+- Fix text query [#33](https://github.com/etalab/cada.data.gouv.fr/pull/33)
+
 
 ## 0.2.1 (2019-07-15)